requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
httplib2==0.22.0
idna==3.7
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pyTelegramBotAPI==4.19.0
requests==2.32.3
rsa==4.9
telebot==0.0.5
uritemplate==4.1.1
urllib3==2.2.1
flask==2.0.1
Werkzeug==2.0
stripe==2.56.0
